+++ src/Form/UserType.php
@@ -8,9 +8,18 @@ use Symfony\Component\Form\Extension\Core\Type\ChoiceType;
 use Symfony\Component\Form\Extension\Core\Type\EmailType;
 use Symfony\Component\Form\Extension\Core\Type\PasswordType;
 use Symfony\Component\Form\FormBuilderInterface;
+use Symfony\Component\HttpFoundation\RequestStack;
+use Symfony\Component\OptionsResolver\OptionsResolver;
 
 class UserType extends AbstractType
 {
+    private $requestStack;
+
+    public function __construct(RequestStack $requestStack)
+    {
+        $this->requestStack = $requestStack;
+    }
+
     public function buildForm(FormBuilderInterface $builder, array $options): void
     {
         $builder
@@ -51,5 +60,20 @@ class UserType extends AbstractType
                 }
             ))
         ;
+    }
+
+    public function configureOptions(OptionsResolver $resolver): void
+    {
+        $ajaxAttrs = ['data-role' => 'modal-form'];
+        $formAttrs = [];
+
+        $attrs = $this->requestStack->getCurrentRequest()->isXmlHttpRequest()
+            ? array_merge($formAttrs, $ajaxAttrs )
+            : $formAttrs
+        ;
+
+        $resolver->setDefaults([
+            'attr' => $attrs,
+        ]);
     }
 }