+++ CHANGELOG.md
@@ -5,6 +5,7 @@
     - support for Czech, English and Russian languages;
     - ability to switch between the available locales;
     - 'Translate to your preferred language' link;
+    - Ajax forms submitting;
 - **Changed**
 - **Deprecated**
 - **Removed**